+++ UI/Widgets/PUIWHScrollBar.cs
@@ -98,7 +98,7 @@ namespace ElementEngine
             var textureCenter = backgroundElCenter == null ? null : (string.IsNullOrWhiteSpace(backgroundElCenter.Value) ? null : AssetManager.LoadTexture2D(backgroundElCenter.Value, preMultiplyAlpha));
 
             var bgWidth = int.Parse(GetXMLElement("Background", "Width").Value);
-            var backgroundTexture = new Texture2D((uint)bgWidth, (uint)textureCenter.Height);
+            var backgroundTexture = new Texture2D(bgWidth, textureCenter.Height);
             
             backgroundTexture.BeginRenderTarget();
             backgroundTexture.RenderTargetClear(RgbaFloat.Clear);
