+++ UI/Widgets/PUIWTextBox.cs
@@ -70,7 +70,7 @@ namespace ElementEngine
             var textureCenter = backgroundElCenter == null ? null : (string.IsNullOrWhiteSpace(backgroundElCenter.Value) ? null : AssetManager.LoadTexture2D(backgroundElCenter.Value, preMultiplyAlpha));
 
             var bgWidth = int.Parse(GetXMLElement("Background", "Width").Value);
-            var backgroundTexture = new Texture2D((uint)bgWidth, (uint)textureCenter.Height);
+            var backgroundTexture = new Texture2D(bgWidth, textureCenter.Height);
             backgroundTexture.BeginRenderTarget();
             backgroundTexture.RenderTargetClear(RgbaFloat.Clear);
 
@@ -138,7 +138,7 @@ namespace ElementEngine
 
             var tSize = _font.MeasureText(_text.Length > 0 ? _text : " ", FontSize);
 
-            _textTexture = new Texture2D((uint)tSize.X, (uint)tSize.Y);
+            _textTexture = new Texture2D((int)tSize.X, (int)tSize.Y);
             _textTexture.BeginRenderTarget();
             _textTexture.RenderTargetClear(RgbaFloat.Clear);
